refactor: split GameFilter and GameSorter components from game page
