+++ components/pages/index.ts
@@ -4,6 +4,8 @@ export * from './connectWallet'
 export * from './filterGroup'
 export * from './gameCarousel'
 export * from './gameCell'
+export * from './gameFilter'
+export * from './gameSorter'
 export * from './loginMethodChooser'
 export * from './pageCard'
 export * from './relatedTags'
