+++ components/pages/sortOptions.tsx
@@ -1,10 +1,11 @@
 import styled from '@emotion/styled'
 import clsx from 'clsx'
+import { SortIcon } from 'components/icons'
 import { SortOptionsContext } from 'context'
+import { useBuildHref } from 'hooks'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import React, { useContext } from 'react'
-import { buildQuerySting } from 'utils'
 
 export declare interface SortOptionsProps {
   sortKey: string
@@ -34,14 +35,14 @@ export function SortOptions({
     padding-bottom: 7px;
     white-space: nowrap;
     display: flex;
+    align-items: center;
   `
-  const Icon = styled.span`
-    & > svg {
-      display: inline-block;
-      vertical-align: middle;
-      margin-right: 10px;
-      color: var(--w3itch-text3);
-    }
+  const StyledSortIcon = styled(SortIcon)`
+    display: inline-block;
+    vertical-align: middle;
+    margin-right: 10px;
+    color: var(--w3itch-text3);
+    font-size: 16px;
   `
   const Sorts = styled.ul`
     list-style: none;
@@ -55,20 +56,7 @@ export function SortOptions({
   return (
     <Container className={className}>
       <Label>
-        <Icon>
-          <svg
-            className="svgicon icon_sort"
-            width="16"
-            height="17"
-            fill="currentColor"
-            role="img"
-            version="1.1"
-            viewBox="0 0 455 488"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path d="m304 392v48c0 4.5-3.5 8-8 8h-64c-4.5 0-8-3.5-8-8v-48c0-4.5 3.5-8 8-8h64c4.5 0 8 3.5 8 8zm-120-32c0 2.25-1 4.25-2.5 6l-79.75 79.75c-1.75 1.5-3.75 2.25-5.75 2.25s-4-0.75-5.75-2.25l-80-80c-2.25-2.5-3-5.75-1.75-8.75s4.25-5 7.5-5h48v-344c0-4.5 3.5-8 8-8h48c4.5 0 8 3.5 8 8v344h48c4.5 0 8 3.5 8 8zm168-96v48c0 4.5-3.5 8-8 8h-112c-4.5 0-8-3.5-8-8v-48c0-4.5 3.5-8 8-8h112c4.5 0 8 3.5 8 8zm48-128v48c0 4.5-3.5 8-8 8h-160c-4.5 0-8-3.5-8-8v-48c0-4.5 3.5-8 8-8h160c4.5 0 8 3.5 8 8zm48-128v48c0 4.5-3.5 8-8 8h-208c-4.5 0-8-3.5-8-8v-48c0-4.5 3.5-8 8-8h208c4.5 0 8 3.5 8 8z" />
-          </svg>
-        </Icon>
+        <StyledSortIcon className="svgicon icon_sort" />
         <div>Sort by</div>
       </Label>
       <Sorts>
@@ -111,11 +99,8 @@ export function SortOptionItem({ name, value }: SortOptionItemProps) {
   `
   const context = useContext(SortOptionsContext)
   const router = useRouter()
-  const href = `${router.route}${buildQuerySting(
-    context.sortKey,
-    value,
-    router.query as Record<string, string>
-  )}`
+  const { buildHref } = useBuildHref()
+  const href = buildHref(context.sortKey, value)
   const isHref = router.asPath === href
 
   return (
